+++ plm-bom-configuration-infrastructure/src/main/java/com/nio/ngfs/plm/bom/configuration/infrastructure/repository/entity/BomsConfigurationRuleGroupEntity.java
@@ -34,7 +34,7 @@ public class BomsConfigurationRuleGroupEntity extends BaseEntity {
     /**
      * 创建Rule的目的，1-（Sales —> Eng），2-（Sales —> Sales），3-（Sales <—> Sales），4-（Sales X Sales），5-（Base Vehicle —> Sales）
      */
-    private Byte purpose;
+    private Integer purpose;
 
     /**
      * Rule适用范围
@@ -49,17 +49,17 @@ public class BomsConfigurationRuleGroupEntity extends BaseEntity {
     @Override
     public String toString() {
         return "BomsConfigurationRuleGroup{" +
-            "id = " + getId() +
-            ", chineseName = " + chineseName +
-            ", displayName = " + displayName +
-            ", purpose = " + purpose +
-            ", definedBy = " + definedBy +
-            ", description = " + description +
-            ", createUser = " + getCreateUser() +
-            ", updateUser = " + getUpdateUser() +
-            ", createTime = " + getCreateTime() +
-            ", updateTime = " + getUpdateTime() +
-            ", delFlag = " + getDelFlag() +
-        "}";
+                "id = " + getId() +
+                ", chineseName = " + chineseName +
+                ", displayName = " + displayName +
+                ", purpose = " + purpose +
+                ", definedBy = " + definedBy +
+                ", description = " + description +
+                ", createUser = " + getCreateUser() +
+                ", updateUser = " + getUpdateUser() +
+                ", createTime = " + getCreateTime() +
+                ", updateTime = " + getUpdateTime() +
+                ", delFlag = " + getDelFlag() +
+                "}";
     }
 }
